Reduce Databases to single type parameter
