+++ packages/truffle-db/src/workspace/index.ts
@@ -1,9 +1,13 @@
 import path from "path";
 
 export { schema } from "./schema";
-import { FSDatabases, MemoryDatabases, SqliteDatabases } from "./pouch";
-import { WorkspaceDatabases } from "./types";
-import { definitions } from "./definitions";
+import {
+  Databases,
+  FSDatabases,
+  MemoryDatabases,
+  SqliteDatabases
+} from "./pouch";
+import { WorkspaceCollections, definitions } from "./definitions";
 
 export interface WorkspaceConfig {
   workingDirectory?: string;
@@ -13,16 +17,8 @@ export interface WorkspaceConfig {
   };
 }
 
-const getDefaultFSAdapterSettings = workingDirectory => ({
-  directory: path.join(workingDirectory, ".db", "json")
-});
-
-const getDefaultSqliteAdapterSettings = workingDirectory => ({
-  directory: path.join(workingDirectory, ".db", "sqlite")
-});
-
 export class Workspace {
-  public databases: WorkspaceDatabases;
+  public databases: Databases<WorkspaceCollections>;
 
   constructor({
     workingDirectory,
@@ -161,3 +157,11 @@ export class Workspace {
     return contracts.map(({ name }) => name);
   }
 }
+
+const getDefaultFSAdapterSettings = workingDirectory => ({
+  directory: path.join(workingDirectory, ".db", "json")
+});
+
+const getDefaultSqliteAdapterSettings = workingDirectory => ({
+  directory: path.join(workingDirectory, ".db", "sqlite")
+});
