+++ packages/truffle-db/src/workspace/pouch/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
+export * from "./databases";
 export * from "./fs";
 export * from "./memory";
 export * from "./sqlite";
